Make search bar and results more accessible
Added information on how to navigate to results in the aria-label
attribute on the search bar

Added a new field to English internationalization with instructions on
how to navigate to search results. This field will appear via the
aria-label to screen reader users.

Added an enclosing div informing screen readers to update when the
search results populate.

Signed-off-by: Patrick Smyth <patrick@iotaschool.com>