Add back descriptive content on cert issuing
As part of doc enhancements in b57dc161, the description in step 7 was
removed. Although the final step is simple and the diagram and section
header are clear, it is jarring for the final step to be inconsistent
with the others by having no text body. It is moreover unfriendly to
screen readers since there is no context for the diagram. This change
adds the brief description back for the sake of readability.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>