gitsign.md: Discourage users from installing an old version
While on a mission to quickly get `gitsign` set up, I admit that I glazed
over the note about going to the releases page to find the latest version
and just copy/pasted the `wget`/`dpkg -i` steps. Well, that left me with
an ancient version of `gitsign`, which I didn't realize until I tried to
make it work with `gitsign-credential-cache.

Discourage this behavior by omitting the version from the instructions,
forcing the user to go look at the releases page.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>